+++ src/views/channels/closedChannels/ClosedCard.tsx
@@ -11,7 +11,7 @@ import {
   getTransactionLink,
   renderLine,
   getNodeLink,
-} from '../../../components/generic/Helpers';
+} from '../../../components/generic/hhelpers';
 import styled from 'styled-components';
 import { Price } from '../../../components/price/Price';
