+++ src/views/tools/backups/Backups.tsx
@@ -13,7 +13,10 @@ import { DownloadBackups } from './DownloadBackups';
 import { VerifyBackups } from './VerifyBackups';
 import { RecoverFunds } from './RecoverFunds';
 import { AdminSwitch } from '../../../components/adminSwitch/AdminSwitch';
-import { getDateDif, getFormatDate } from '../../../components/generic/Helpers';
+import {
+  getDateDif,
+  getFormatDate,
+} from '../../../components/generic/hhelpers';
 
 export const BackupsView = () => {
   const [lastDate, setLastDate] = useState('');
